ENG-2031: Adding logging and dbg statements for debugging
This has helped narrow down the problem. Logging confirms that the
processor does process all messages. My current hunch is that rate
limiting is getting close incorrectly in some cases when we still have
messages. Those messages eventually timeout and the channel gets closed.

The point against this is that I don't think it matches what I'm seeing
in traces in Datadog. It could be that I'm misinterpreting the traces
though.