ENG-2031: Simplified bug fix
When the next message is too large to send, this version attempts to
set subtract the remaining demand from the rate limit to set it to 0.
This triggers the rate limit to close correctly.

I'm still seeing the Bandwidth SMS queue stuck locally, but the T-Mobile
queue is flowing correctly.